project-echo-package/requirements.txt
numpy
ffmpeg-python==0.2.0
keras
tensorflow
tensorflow-io
tfimm
pydub
os
